ci: refactor long gate files below limit
